requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
azure-ai-documentintelligence==1.0.2
azure-core==1.36.0
azure-identity==1.25.1
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
defusedxml==0.7.1
distro==1.9.0
elevenlabs==2.20.1
et_xmlfile==2.0.0
fastapi==0.119.1
flatbuffers==25.9.23
griffe==1.14.0
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
jiter==0.11.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
magika==0.6.2
mammoth==1.10.0
markdownify==1.2.0
mcp==1.18.0
mem0ai==1.0.0
mangum==0.17.0
mpmath==1.3.0
msal==1.34.0
msal-extensions==1.3.1
numpy==2.3.4
olefile==0.47
ollama==0.6.0
onnxruntime==1.23.2
openai==2.6.0
openai-agents==0.4.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
passlib[bcrypt]==1.7.4
pypdf==5.1.0
pillow==12.0.0
portalocker==3.2.0
posthog==6.7.8
protobuf==5.29.5
pycparser==2.23
pydantic==2.12.3
pydantic-settings==2.11.0
pydantic_core==2.41.4
pydub==0.25.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pymongo==4.10.1
qdrant-client==1.15.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SpeechRecognition==3.14.3
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
tqdm==4.67.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
xlrd==2.0.2
xlsxwriter==3.2.9
youtube-transcript-api==1.0.3
tiktoken==0.7.0